+++ src/components/page/page.js
@@ -4,22 +4,14 @@ import classNames from 'classnames';
 
 export const EuiPage = ({ children, className, restrictWidth, style, ...rest }) => {
   let widthClassname;
+  let newStyle;
 
   if (restrictWidth === true) {
     widthClassname = 'euiPage--restrictWidth-default';
-  } else if (restrictWidth === false) {
-    widthClassname = 'euiPage--widthIsNotRestricted';
-  } else {
+  } else if (restrictWidth !== false) {
     widthClassname = 'euiPage--restrictWidth-custom';
-
-    // if style has been passed as a prop, add to it
-    if (style) {
-      style.maxWidth = `${restrictWidth}px`;
-    }
-    // otherwise create a new object
-    else {
-      style = { maxWidth: `${restrictWidth}px` };
-    }
+    const value = typeof maxWidth === 'number' ? `${restrictWidth}px` : restrictWidth;
+    newStyle = { ...style, maxWidth: value };
   }
 
   const classes = classNames(
@@ -31,7 +23,7 @@ export const EuiPage = ({ children, className, restrictWidth, style, ...rest }) 
   return (
     <div
       className={classes}
-      style={style}
+      style={newStyle || style}
       {...rest}
     >
       {children}
@@ -47,11 +39,13 @@ EuiPage.propTypes = {
    * Sets the max-width of the page,
    * set to `true` to use the default size,
    * set to `false` to not restrict the width,
-   * set to a number for a custom width.
+   * set to a number for a custom width in px,
+   * set to a string for a custom width in custom measurement.
    */
   restrictWidth: PropTypes.oneOfType([
     PropTypes.bool,
-    PropTypes.number
+    PropTypes.number,
+    PropTypes.string,
   ]),
 };