+++ src/components/page/page_body/page_body.test.js
@@ -6,11 +6,28 @@ import { EuiPageBody } from './page_body';
 
 describe('EuiPageBody', () => {
   test('is rendered', () => {
-    const component = render(
-      <EuiPageBody {...requiredProps} />
-    );
+    const component = render(<EuiPageBody {...requiredProps} />);
 
-    expect(component)
-      .toMatchSnapshot();
+    expect(component).toMatchSnapshot();
+  });
+
+  describe('restrict width', () => {
+    test('can be set to a default', () => {
+      const component = render(<EuiPageBody {...requiredProps} restrictWidth={true} />);
+
+      expect(component).toMatchSnapshot();
+    });
+
+    test('can be set to a custom number', () => {
+      const component = render(<EuiPageBody {...requiredProps} restrictWidth={1024} />);
+
+      expect(component).toMatchSnapshot();
+    });
+
+    test('can be set to a custom value and measurement', () => {
+      const component = render(<EuiPageBody {...requiredProps} restrictWidth="24rem" />);
+
+      expect(component).toMatchSnapshot();
+    });
   });
 });